Trigonometry calculator
This Python script can calculate the sine, cossine and tangent of a desired angle (degree), via Leibniz series. It has a degree-to-radian converter, factorial function, pi calculator and the sine, cossine and tangent functions.
It's worth noting that the factorial is possible until 175!. After that, it'll exceed Python's numeric limit. So the calculations will become unprecise in angles more than 720º.